cookbooks/factotum-agent-py/requirements.txt
swarmkit>=0.1.27
python-dotenv
rich
prompt_toolkit>=3.0.0
